Changed module path to module name
FBS specified the entry point using the entry point file path.
This worked well when the code was run in place but became problematic when it was preferred to use the installed version instead.
This changes the definition of the entry point to the module name which can be used to import the module and find the module path that is being used by python.
If it cannot be found, the python_path variable is added to sys.path allowing the source version to be used.