Added ability to get version from the package
FBS had the version number hard coded into the build system and injected it into the program via a hook.
This felt rather janky when there are many existing tools to manage the version.
This change adds the ability to pull the version number from an attribute in the package itself.